ipa: Add debug controls to the agc algorithm
Add a few (arbitrary) debug metadata to the agc algorithm. This shows
how easy it is to add debug metadata to an ipa even for classes that
don't have direct access to the metadata list or the context like
AgcMeanLuminance. The control_ids_debug.yaml was autogenerated by
calling utils/gen-debug-controls.py

Signed-off-by: Stefan Klug <stefan.klug@ideasonboard.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>