libcamera: Add mailbox template helper
MailBox is designed to be used to shares data, usually a frame buffer, among
tasks. A Mailbox can store a data which the producer task should set into,
and multiple consumer tasks could read the data from it. The data could
be set only once, i.e., there should be only one producer. All tasks shares
the Mailbox with share_ptr and the data it contains should be recycled when
all tasks are finished.

Signed-off-by: Han-Lin Chen <hanlinchen@chromium.org>
Co-developed-by: Harvey Yang <chenghaoyang@chromium.org>
Signed-off-by: Harvey Yang <chenghaoyang@chromium.org>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>