libcamera: Add InfoFrame implememtation
InfoFrame consists of an InfoFrame class which is used as an extension
of FrameBuffer, with extra information of pixel format, size, and memory
address if it's mmapped.

InfoFramePool is introduced to handle InfoFrames more easily, which supports
a pool implementation of InfoFrames, creates frames by DmaBufAllocator, and
mmap/munmap all frames of the pool.

Signed-off-by: Han-Lin Chen <hanlinchen@chromium.org>
Co-developed-by: Harvey Yang <chenghaoyang@chromium.org>
Signed-off-by: Harvey Yang <chenghaoyang@chromium.org>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>