libcamera: controls: Add ControlList::merge overload that moves entries
When merging two control lists, copies of the entries are made. Add a
overload of the merge function that accepts a non-const source and
provides a merge without copy. This was deemed important in the light of
debug metadata were the data should not be copied more often than
necessary.

This change does not break the API, but it changes the behavior of
applications because the source list now gets modified if it is
non-const (which might be a common case).

Signed-off-by: Stefan Klug <stefan.klug@ideasonboard.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>