Fix copyright year
without this patch, openSUSE's debugsource package would vary over time,
breaking reproducible builds.
Also a copyright of 2025 is factually wrong,
because nobody did anything copyright-worthy in that year.
If updating the year manually is too hard (it is optional anyway),
using the mtime of input file(s) could work as well.

By moving the year to tracepoints.h.in it can easily be updated
along with the content.

The diff was:
/usr/src/debug/libcamera-0.3.1/x86_64-suse-linux/include/libcamera/internal/tracepoints.h       2024-08-20 09:37:46.000000000 +0200
@@ -1,6 +1,6 @@
 /* SPDX-License-Identifier: LGPL-2.1-or-later */
 /*
- * Copyright (C) 2024, Google Inc.
+ * Copyright (C) 2025, Google Inc.

Signed-off-by: Bernhard M. Wiedemann <bwiedemann@suse.de>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>