rkisp1: synchronize sensor controls with IPA
When IPA AGC active state is initialized in configure() method, it may
fall out of sync with actual sensor controls and initial values in the
delayed controls. Fix this by sending updated controls to the pipeline
handler on first start() after configure(). Also reset delayedCtrls_
on start to make it capture the updated control values.

Signed-off-by: Mikhail Rudenko <mike.rudenko@gmail.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>