libcamera: rkisp1: Enable the dewarper unconditionally
In configure() and in the future in generateConfiguration() the
calculated stream sizes and crop rectangles depend on the dewarper being
used or not. It is therefore not possible to postpone that decision
until the dewarper gets configured. Enable the dewarper unconditionally
if it is found and the stream type is not RAW.

Signed-off-by: Stefan Klug <stefan.klug@ideasonboard.com>
Signed-off-by: Jacopo Mondi <jacopo.mondi@ideasonboard.com>