libcamera: software_isp: Stop clearing context config and state again
This partly reverts commit 41e3d61c, removing parts that had unintended
side effects as the intention for the commit was purely to fix crashes.

Clearing the configuration turned out to be problematic as some values such
as configuration.black.level only get on initialization and thus were never
used.

Clearing the activeState resulted in additional, arguably undesired churn,
very noticable when switching back and forth between cameras. Whether this
is desirable is AFAIK a matter of taste/policy and shouldn't have been done
as part of a crash fix.

Fixes: 41e3d61c ("libcamera: software_isp: Clear IPA context on configure and stop")
Signed-off-by: Robert Mader <robert.mader@collabora.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>