ipa: rkisp1: use active state in setControls
The results of AGC algorithm currently travel a long way before being
actually applied. Let's consider the common case when 4 requests are
used and frames 0-3 are queued before start(). AGC is first run when
frame 0 stats are ready, and exposure/gain are saved in
activeState. Then, when frame 4 is queued, they are stored in
corresponding frameContext. Four frames later, frame 4 is captured,
and after processing the stats buffer, exposure/gain are extracted
from frameContext and queued to delayedControls in setControls(). On
frame 5, delayedControls apply them. Assuming a control delay of 2, the
settings will become effective for frame 7. So, it takes 7 frames for
the AGC algorithm to get feedback. This results in suboptimal
convergence rate.

If we instead use just computed exposure/gain from the activeState in
setControls, the delay is reduced from 7 frames to 3 frames. Tests
on OV4689 sensor show faster convergence and no unwanted side effects.

Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>