libcamera: Rename "shutter speed" to "exposure time"
The terms "shutter" and "shutter speed" are used through libcamera to
mean "exposure time". This is confusing, both due to "speed" being used
as "time" while it should be the inverse (i.e. a maximum speed should
correspond to the minimum time), and due to "shutter speed" and
"exposure time" being used in different places with the same meaning.

To improve clarity of the code base and the documentation, use "exposure
time" consistently to replace "shutter speed".

I'm sending this patch as an RFC to gather opinions before addressing a
few remaining issues, regarding which I'd appreciate feedback as well:

- The tuning data for the AgcMeanLuminance helper class still uses a
  "shutter" YAML element. I would ideally like to rename that too.

- The ExposureModeHelper::splitExposure() function used to document that
  it splits "exposure time into shutter time and gain". I have reworded
  that to "split exposure into exposure time and gain", but I'm not
  entirely happy with the result as "exposure" has a defined meaning in
  photography (see https://en.wikipedia.org/wiki/Exposure_(photography))
  that is not expressed as a duration. It seems we need a better term
  here.

- I haven't touched the Raspberry Pi IPA module. David, Naush, would you
  accept a rename there too ?

Signed-off-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>