libcamera: rkisp1: Rationalize IPA and handlers names
The names used by the IPA interface and the names used for buffer
completions handlers for the RkISP1 clash in the use of the term "buffer".

For example the video device buffer completion handler is called
"bufferReady" and the IPA event to ask the IPA to compute parameters is
called "fillParamsBuffers". This makes it hard to recognize which
function handles video device completion signals and which ones handle
the IPA interface events.

Rationalize the naming scheme in the IPA interface function and events
and the signal handlers in the RkISP1 support, according to the
following table. Remove the name "buffer" from the IPA interface events
and events handler and reserve it for the buffer completion handlers.
Rename the IPA interface events and function to use the 'params' and
'stats' names and make the signal and the even names specular (ie
'computeParams' : 'paramsComputed')

IPA Interface:

- fillParamsBuffer -> computeParams   [FUNCTION]
- paramsBufferReady -> paramsComputed [EVENT]

- processStatsBuffer -> processStats  [FUNCTION]
- metadataReady -> statsProcessed     [EVENT]

Pipeline handler:

- bufferReady -> videoBufferReady     [BUFFER HANDLER]
- paramReady -> paramsBufferReady     [BUFFER HANDLER]
- statReady -> statsBufferReady       [BUFFER HANDLER]
- paramFilled -> paramsComputed       [IPA EVENT HANDLER]
- metadataReady -> statsProcessed     [IPA EVENT HANDLER]

Cosmetic change only, no functional changes intended.

Signed-off-by: Jacopo Mondi <jacopo.mondi@ideasonboard.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>