ipa: rkisp1: Have algos initialize FrameContext
The RkISP1 AGC algorithms assumes the metering mode to be
"MeteringMatrix" and pre-fill an array of weights associated
with it stored in meteringModes_[MeteringMatrix] when intializing
the algorithm in parseMeteringModes().

It laters fetches the weights when computing parameters using the
FrameContext.agc.meteringMode as index of the meteringModes_ array.

When a Request gets queued to the algorithm, the
FrameContext.agc.meteringMode index is populated from the algorithm's
active state, and optionally updated if the application has specified
a different metering mode.

In case of Request underrun however, a non-initialized FrameContext
gets provided to the algorithm, causing an (unhandled) exception when
accessing the meteringModes_ array.

To handle the situation when a Request underrun occours and let
algorithms initialize a FrameContext to safe defaults:

- Add an 'underrun' flag to FrameContext
- Add an 'initFrameContext()' function to RkISP1 algorithms
- If a FrameContext is get() before being allocated, pass it through
  the algorithms' initFrameContext() to safely initialize it

Signed-off-by: Jacopo Mondi <jacopo.mondi@ideasonboard.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>