libcamera: camera_sensor_properties: Add sensor control delays
Add properties covering the sensor control application delays to both
the list of control values and the static CameraSensorProperties
definitions. The values used are the defaults that're in use across
the library, with deviations from that taken from Raspberry Pi's
CamHelper class definitions.

Initialise the properties from the static database during the same
CameraSensor::initStaticProperties() function that currently handles
the UnitCellSize.

Signed-off-by: Daniel Scally <dan.scally@ideasonboard.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>