ipa: raspberry: Port to the new AEGC controls
The newly introduced controls to drive the AEGC algorithm allow
to control the computation of the exposure time and analogue gain
separately.

The RPi AEGC implementation already computes the shutter and gain
values separately but does not expose separate functions to control
them.

Augment the AgcAlgorithm interface to allow pausing/resuming the shutter
and gain automatic computations separately and plumb them to the newly
introduced controls.

Add safety checks to ignore ExposureTime and AnalogueGain values if the
algorithms are not paused, and report the correct AeState value by
checking if both algorithms have been paused or if they have converged.

Signed-off-by: Jacopo Mondi <jacopo@jmondi.org>
Signed-off-by: Paul Elder <paul.elder@ideasonboard.com>
Reviewed-by: Paul Elder <paul.elder@ideasonboard.com>
Reviewed-by: Stefan Klug <stefan.klug@ideasonboard.com>

---
Changes in v8.1:
- fix the racy check in handling the auto/manual controls and the manual
  control value control

Changes in v8:
- add auto as default value for ExposureMode and AnalogueGainMode
- add todo for control processing order

No change in v7

No change in v6

No change in v5

Changes in v4:
- s/shutter/exposure/

Changes in v3:
- recovered from 2-year-old bitrot