gstreamer: Read correct colorimetry from caps into the stream configuration
When libcamerasrc is negotiating with downstream element, it first
extracts colorimetry field from downstream supported caps, then set
this colorimetry to its stream configuration and propagates the
colorimetry downstream.

Currently libamerasrc only considers the case there is one colorimetry
in colorimetry field of downstream caps. But the issue is that
downstream caps may report a list of supported colorimetry, which
causes libcamerasrc to set unknown colorimetry to stream configuration
and negotiate fail with downstream element.

In order to fix the issue, get the first string if colorimetry field
holds string list to set it to stream configuration.

Signed-off-by: Hou Qi <qi.hou@nxp.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>