libcamera: libipa: camera_sensor: Add Sony IMX462 sensor properties
Hi,
This patch-set comes as part of an ongoing effort to support Sony IMX327 and IMX462 sensors better in the linux kernel.
For the kernel following changes are proposed:
https://github.com/raspberrypi/linux/pull/5859

After those changes have been applied, the 2 Starvis sensors will no longer be identied by libcamera as a IMX290 and therefore require this patch-set to work correctly.

Signed-off-by: Geoffrey Van Landeghem <geoffrey.vl@gmail.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>