libcamera: Copy matrix class from libipa to libcamera
In preparation to moving the Matrix implementation from libipa to
libcamera copy the corresponding files to the new location. The files
are copied without modification to make upcoming integration commits
easier to understand. The new files are not included in the build and
therefore have no negative side effects on the build.

Signed-off-by: Stefan Klug <stefan.klug@ideasonboard.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>