libcamera: simple: Fix raw output
Simple pipeline was used to introduce software ISP.  If software ISP is
enabled, the pipeline always debayers the input and doesn't produce raw
output anymore, even when requested.

This patch fixes the problem and allows producing raw output if it is
requested, by specifying raw stream role.  When raw output is requested,
only single stream (raw or processed) is allowed.  This is not
necessarily an inherent property, it just simplifies the implementation
and is sufficient for the current needs.

If raw stream is requested, it means raw as it is and no pixel format or
size conversions are permitted.  This is not only logical but also
simplifying the changes because we don't have to check for raw when
conversion is used.

The patch must deal with the fact that software ISP and its output
configurations are arranged before we know whether a raw or a processed
stream will be requested.  This means that if software ISP is enabled
then it is always initialized and we store raw output configurations
together with software ISP ones.  We filter output configurations later
as needed.

The raw output may not be useful without exposure and gain adjustments,
it can be all black (or just noisy).  This will be handled in a separate
patch introducing manual gain and exposure controls.

Signed-off-by: Milan Zamazal <mzamazal@redhat.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>